backend/requirements.txt
Django==4.2.24
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.10
django-redis==6.0.0
redis==6.4.0
celery==5.5.3
requests==2.32.5